Change: BottlePrice property in the ReleasesObject class is now a ulong

+++ src/Whiskey-Tycoon.lib/JSONObjects/ReleasesObject.cs
@@ -12,7 +12,7 @@
 
         public int ReleaseQuarter { get; set; }
 
-        public float BottlePrice { get; set; }
+        public ulong BottlePrice { get; set; }
 
         public ulong BottlesSold { get; set; }
